test(fixtures): refactor
